// packages/super-json-api-client/src/models/CoreTypes_EventHistoryItem.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CoreTypes_EventHistoryItem = {
    action: string;
    amount: string;
    transaction_version: string;
    transaction_block_height: number;
    broker_id: string;
    portfolio_address: string;
    created: string;
};