// packages/super-json-api-client/src/models/CoreTypes_EventTransaction.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CoreTypes_EventTransaction = {
    id: number;
    created: number;
    broker_id: string;
    transaction_version: string;
    portfolio_address: string;
    action: string;
    amount: string;
};