// packages/super-json-api-client/src/models/CoreTypes_PointsHistoryTotals.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CoreTypes_PointsHistoryTotals = {
    id: number;
    address: string;
    collateraldailydollars: number;
    liabilitydailydollars: number;
    dailypoints: number;
    runningtotalvalue: number;
    runningtotalpoints: number;
    runningtotallendingpoints: number;
    runningtotalborrowingpoints: number;
    ts: string;
    multiplier: number;
    ranking: number;
    ranking_total: number;
    referal_points: number;
    referal_multiplier: number;
    referal_count: string;
    runningtotalreferalpoints: number;
};